Add search to sidebar
